refactor(Const): 로또 구매 단위 타입 int -> long 변경

+++ src/main/java/lotto/util/Const.java
@@ -5,5 +5,5 @@ public class Const {
     public static final int START_NUMBER = 1;
     public static final int END_NUMBER = 45;
     public static final int LOTTO_SIZE = 6;
-    public static final int LOTTO_PRICE_UNIT = 1000;
+    public static final long LOTTO_PRICE_UNIT = 1000L;
 }
